#14 changed type assignment and renamed utils.ts to formatTime.ts
