+++ src/utils/formatTime.ts
